+++ policyengine-client/src/countries/country.jsx
@@ -55,7 +55,7 @@ export default class Country {
         let value;
         for (let variable of this.inputVariables) {
             metadata = this.variables[variable];
-            if(!metadata) {
+            if (!metadata) {
                 throw new Error(`Failed to load ${variable}.`)
             }
             value = metadata.valueType === "Enum" ? metadata.defaultValue.key : metadata.defaultValue;
@@ -68,7 +68,7 @@ export default class Country {
         }
         for (let variable of this.outputVariables) {
             metadata = this.variables[variable];
-            if(!metadata) {
+            if (!metadata) {
                 throw new Error(`Failed to load ${variable}.`)
             }
             entity = this.entities[metadata.entity];