+++ policyengine/countries/uk/default_reform/reform.py
@@ -16,6 +16,7 @@ import pandas as pd
 import warnings
 
 warnings.filterwarnings("ignore")
+baseline_parameters = CountryTaxBenefitSystem().parameters
 
 
 def add_extra_band(parameters: ParameterNode) -> ParameterNode:
@@ -218,7 +219,7 @@ def create_default_reform() -> ReformType:
     class adjusted_net_income(Variable):
         value_type = float
         entity = Person
-        label = u"Taxable income after tax reliefs and before allowances"
+        label = "Taxable income after tax reliefs and before allowances"
         definition_period = YEAR
         reference = "Income Tax Act 2007 s. 23"
 
@@ -245,7 +246,7 @@ def create_default_reform() -> ReformType:
     class UC_earned_income(Variable):
         value_type = float
         entity = BenUnit
-        label = u"Universal Credit earned income (after disregards and tax)"
+        label = "Universal Credit earned income (after disregards and tax)"
         definition_period = YEAR
 
         def formula(benunit, period, parameters):
@@ -269,7 +270,7 @@ def create_default_reform() -> ReformType:
     class tax_credits_applicable_income(Variable):
         value_type = float
         entity = BenUnit
-        label = u"Applicable income for Tax Credits"
+        label = "Applicable income for Tax Credits"
         definition_period = YEAR
         reference = "The Tax Credits (Definition and Calculation of Income) Regulations 2002 s. 3"
 
@@ -303,7 +304,7 @@ def create_default_reform() -> ReformType:
     class guarantee_credit_applicable_income(Variable):
         value_type = float
         entity = BenUnit
-        label = u"Applicable income for Pension Credit"
+        label = "Applicable income for Pension Credit"
         definition_period = YEAR
 
         def formula(benunit, period, parameters):
@@ -344,7 +345,7 @@ def create_default_reform() -> ReformType:
     class JSA_income_applicable_income(Variable):
         value_type = float
         entity = BenUnit
-        label = u"Relevant income for JSA (income-based) means test"
+        label = "Relevant income for JSA (income-based) means test"
         definition_period = YEAR
 
         def formula(benunit, period, parameters):
@@ -390,7 +391,7 @@ def create_default_reform() -> ReformType:
     class income_support_applicable_income(Variable):
         value_type = float
         entity = BenUnit
-        label = u"Relevant income for Income Support means test"
+        label = "Relevant income for Income Support means test"
         definition_period = YEAR
 
         def formula(benunit, period, parameters):
@@ -436,7 +437,7 @@ def create_default_reform() -> ReformType:
     class housing_benefit_applicable_income(Variable):
         value_type = float
         entity = BenUnit
-        label = u"Relevant income for Housing Benefit means test"
+        label = "Relevant income for Housing Benefit means test"
         definition_period = YEAR
 
         def formula(benunit, period, parameters):
@@ -609,12 +610,22 @@ def create_default_reform() -> ReformType:
             person = household.members
             tax_band = person("tax_band", period)
             bands = tax_band.possible_values
-
-            return household.any(
+            has_basic_rate_taxpayers = household.any(
                 (tax_band == bands.BASIC)
                 | (tax_band == bands.STARTER)
                 | (tax_band == bands.INTERMEDIATE)
-            ) & ~household("smf_benefit_payment_eligible", period)
+            )
+            has_higher_rate_taxpayers = household.any(
+                (tax_band == bands.HIGHER) | (tax_band == bands.ADDITIONAL)
+            )
+            benefit_payment_eligible = household(
+                "smf_benefit_payment_eligible", period
+            )
+            return (
+                has_basic_rate_taxpayers
+                & ~has_higher_rate_taxpayers
+                & ~benefit_payment_eligible
+            )
 
     class smf_benefit_cash_payment(Variable):
         entity = Household
@@ -625,7 +636,6 @@ def create_default_reform() -> ReformType:
         unit = "currency-GBP"
 
         def formula(household, period, parameters):
-
             rate = parameters(period).reforms.smf_cash_payment.benefit
             return household("smf_benefit_payment_eligible", period) * rate
 
@@ -641,6 +651,19 @@ def create_default_reform() -> ReformType:
             eligible = household("smf_tax_payment_eligible", period)
             rate = parameters(period).reforms.smf_cash_payment.tax
             return eligible * rate
+
+    class personal_allowance(baseline_variables["personal_allowance"]):
+        def formula(person, period, parameters):
+            default_pa = baseline_variables["personal_allowance"].formula(
+                person, period, parameters
+            )
+            is_sp_age = person("is_SP_age", period)
+            baseline_pa = baseline_variables["personal_allowance"].formula(
+                person, period, baseline_parameters
+            )
+            if parameters(period).reforms.misc.exempt_seniors_from_PA_changes:
+                return where(is_sp_age, baseline_pa, default_pa)
+            return default_pa
 
     class default_reform(Reform):
         def apply(self):
@@ -673,5 +696,6 @@ def create_default_reform() -> ReformType:
             self.update_variable(smf_benefit_cash_payment)
             self.update_variable(smf_tax_cash_payment)
             self.update_variable(household_benefits)
+            self.update_variable(personal_allowance)
 
     return default_reform